% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{code_tables}
\alias{code_tables}
\alias{get_code_table}
\alias{get_code_tables}
\title{Get code table data}
\usage{
get_code_table(code_table, name_col, value_col, database)

get_code_tables(code_tables, database)
}
\arguments{
\item{code_table}{A code table name (database table name)}

\item{name_col}{Cold table field name with codes}

\item{value_col}{Cold table field name with values}

\item{database}{Database logical name}

\item{code_tables}{A list of code tables with elements: code_table, name_col,
value_col}
}
\value{
Returns a code table as a list of values
}
\description{
Get code table data

Returns code table as a list of values
}

