% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datamodel.R
\name{display_graph}
\alias{display_graph}
\title{Display graph with \code{DiagrammeR::grViz} (using knitr sizing)}
\usage{
display_graph(diagram = "", engine = "dot", allow_subst = TRUE,
  options = NULL, width = NULL, height = NULL)
}
\arguments{
\item{diagram}{spec for a diagram as either text, filename string, or file connection.}

\item{engine}{string for the Graphviz layout engine}

\item{allow_subst}{a boolean that enables/disables subsitution functionality}

\item{options}{parameters supplied to the htmlwidgets framework}

\item{width}{width}

\item{height}{height}
}
\description{
Display graph with \code{DiagrammeR::grViz} (using knitr sizing)
}

