% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{get_database_data}
\alias{get_database_data}
\title{Get database data}
\usage{
get_database_data(db_name, query)
}
\arguments{
\item{db_name}{Logical database name}

\item{query}{SQL query string}
}
\description{
Get database data
}

