% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datamodel.R
\name{get_db_connection_string}
\alias{get_db_connection_string}
\title{Get DB connecton string}
\usage{
get_db_connection_string(database)
}
\arguments{
\item{database}{Logical name of the database}
}
\value{
Connection string
}
\description{
Get environment variable RMDPC_CON_STRING_[DatabaseName]
}
\keyword{internal}

