% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datamodel.R
\name{get_db_meta}
\alias{get_db_meta}
\title{Get meta data from database}
\usage{
get_db_meta(database = NULL, query_str = NULL, con_string = NULL)
}
\arguments{
\item{database}{Logical name of the database}

\item{query_str}{Query to execute (by default the predefined query is used)}

\item{con_string}{Connection string - by default it is read from
environment variable named RMDOC_CON_STRING_[database]}
}
\value{
A data frame with all columns and related information
}
\description{
Executes SQL query on system tables to get information about database datamodel.
The result is used by \code{\link{get_db_model}} to extract a data model object.
}

