% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datamodel.R
\name{get_db_meta_desc}
\alias{get_db_meta_desc}
\title{Get table descriptions from database}
\usage{
get_db_meta_desc(database = NULL, query_str = NULL, con_string = NULL)
}
\arguments{
\item{database}{Logical name of the database}

\item{query_str}{Query to execute (by default the predefined query is used)}

\item{con_string}{Connection string - by default it is read from
environment variable named RMDOC_CON_STRING_[database]}
}
\value{
A data frame with all columns and related information
}
\description{
Executes SQL query on system tables to get information about tables.
}

