% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datamodel.R
\name{get_db_model}
\alias{get_db_model}
\title{Get data model object from db meta}
\usage{
get_db_model(table_filter = NULL, db_meta = NULL)
}
\arguments{
\item{table_filter}{A list with table names (tables = ...)}

\item{db_meta}{All database meta data (see get_db_meta)}
}
\description{
Get data model object from db meta
}

