% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datamodel.R
\name{get_dm_from_yaml}
\alias{get_dm_from_yaml}
\title{Get data model from yaml definition}
\usage{
get_dm_from_yaml(x, db_meta = NULL)
}
\arguments{
\item{x}{YAML string with data model definition}

\item{db_meta}{Database meta data}
}
\value{
Data model object
}
\description{
Combine tables from YAML definition (table: elements) and
reverse engineered model (extract_to: elements)
}

