% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{get_excel_data}
\alias{get_excel_data}
\title{Get Excel table}
\usage{
get_excel_data(file, skip_empty = TRUE, replace_na = NA)
}
\arguments{
\item{file}{File name}

\item{skip_empty}{Skip empty rows}

\item{replace_na}{Replace NAs with this value}
}
\description{
Get Excel table
}

