% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datamodel.R
\name{get_graph_from_dm}
\alias{get_graph_from_dm}
\title{Get graph from data model}
\usage{
get_graph_from_dm(dm, display_options = NULL)
}
\arguments{
\item{dm}{Datamodel object}

\item{display_options}{Graph display options (see \code{\link{graph_display_options}})}
}
\value{
graphviz dot string
}
\description{
Get graph from data model
}

