% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datamodel.R
\name{get_graph_from_yaml}
\alias{get_graph_from_yaml}
\title{Get graph from YAML}
\usage{
get_graph_from_yaml(x, display_options = NULL, db_meta = NULL)
}
\arguments{
\item{x}{YAML string or .yaml file name}

\item{display_options}{Graph display options (see \code{\link{graph_display_options}})}

\item{db_meta}{Database meta data}
}
\description{
Get graph from YAML
}

