% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datamodel.R
\name{get_table_description}
\alias{get_table_description}
\title{Get Table Description}
\usage{
get_table_description(tables, db_meta = NULL, db_meta_desc = NULL)
}
\arguments{
\item{tables}{Vector of table names}

\item{db_meta}{Database meta data}

\item{db_meta_desc}{Database meta data with table descriptions}
}
\description{
Prints table description markdown
}

