% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datamodel.R
\name{graph_display_options}
\alias{graph_display_options}
\title{Graph display options}
\usage{
graph_display_options(rankdir = "RL", view_type = "all",
  outside_ref = FALSE, show_types = FALSE, show_descriptions = FALSE,
  max_cols = 30)
}
\arguments{
\item{rankdir}{A graphviz rankdir option ("BT", "RL", "TB", "LR")}

\item{view_type}{Can be "all" (by default), "keys_only" or "title_only". It defines the level of details for the table rendering (all columns, only primary and foreign keys or no columns)}

\item{outside_ref}{Include tables only referenced but not specified in diagram definition}

\item{show_types}{Show type description of a column}

\item{show_descriptions}{Show descriptions}

\item{max_cols}{Max columns displayed (when view_type == "all")}
}
\description{
Graph display options used in get_graph_from_yaml or get_graph_from_dm
}

