% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datamodel.R
\name{print_diagram_description}
\alias{print_diagram_description}
\title{Print Diagram Description}
\usage{
print_diagram_description(diagram_file, db_meta = NULL, db_meta_desc = NULL)
}
\arguments{
\item{diagram_file}{A yaml file with diagram definition}

\item{db_meta}{Database meta data}

\item{db_meta_desc}{Database meta data with table descriptions}
}
\description{
Print Diagram Description
}

