% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{run_sql}
\alias{run_sql}
\title{Run SQL}
\usage{
run_sql(sql_file, database, ...)
}
\arguments{
\item{sql_file}{SQL query file}

\item{database}{Logical database name}

\item{...}{query parameters}
}
\description{
Run SQL
}

